# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OMPAY API",
        "description": "API pour le système de paiement mobile MOMPAY",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comptes": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Create a new compte",
                "description": "Create a new compte for a user.",
                "operationId": "d46221922144c672a4a500c7f3755629",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "uuid"
                                    },
                                    "solde": {
                                        "type": "number",
                                        "example": 1000
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "inactif"
                                        ],
                                        "example": "actif"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "client",
                                            "marchand"
                                        ],
                                        "example": "client"
                                    },
                                    "code_marchand": {
                                        "type": "string",
                                        "example": "MARCHAND123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "solde": {
                                                    "type": "number"
                                                },
                                                "status": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/compte/{id}/solde": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Get compte solde by numero_compte",
                "description": "Get the solde of a compte by numero_compte.",
                "operationId": "60d128de95f3b4903fa8275b2004b3d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Numero du compte",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte solde",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "solde": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/compte/{id}/transfert": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Transfer money from compte",
                "description": "Perform a transfer from the compte.",
                "operationId": "9c15e54886bbbd3cf37859b175cc1237",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero",
                                    "montant"
                                ],
                                "properties": {
                                    "numero": {
                                        "type": "string"
                                    },
                                    "montant": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfer successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/compte/{id}/paiement": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Make a payment from compte",
                "description": "Perform a payment from the compte.",
                "operationId": "9e079e4ba6a1f6d9f038df15483ce1d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "montant"
                                ],
                                "properties": {
                                    "telephone": {
                                        "type": "string"
                                    },
                                    "code_marchant": {
                                        "type": "string"
                                    },
                                    "montant": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/compte/{id}/transaction": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Get compte transactions",
                "description": "Get transactions for the compte.",
                "operationId": "9b0f5c2cd4e8dbf0cf8a2ccb7aaf5bad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/compte/dashboard": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Get compte dashboard",
                "description": "Get compte dashboard for authenticated user.",
                "operationId": "c481c5e3625dbdfc4e72c602b3a97b24",
                "responses": {
                    "200": {
                        "description": "Dashboard data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "compte": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter your Bearer token"
            },
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter your Bearer token"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}